+++ src/main/java/biz/gabrys/maven/plugins/css/splitter/steadystate/ParserOptions.java
@@ -19,4 +19,6 @@ public interface ParserOptions {
     Standard getStandard();
 
     boolean isStrict();
+
+    boolean isStarHackAllowed();
 }