+++ src/test/java/biz/gabrys/maven/plugins/css/splitter/steadystate/Css30CompatibilityTest.java
@@ -13,7 +13,8 @@ import biz.gabrys.maven.plugins.css.splitter.css.type.UnknownRule;
 
 public final class Css30CompatibilityTest {
 
-    private final ParserOptions options = new ParserOptionsBuilder().withStandard(Standard.VERSION_3_0).withStrict(false).create();
+    private static final Standard STANDARD = Standard.VERSION_3_0;
+    private static final ParserOptions options = new ParserOptionsBuilder().withStandard(STANDARD).withStrict(false).create();
 
     @Test(expected = ParserException.class)
     public void parse_mediaStoresCharsetRule_throwsException() {
@@ -39,12 +40,15 @@ public final class Css30CompatibilityTest {
         parser.parse("@media all { @import 'file.css'; }", options);
     }
 
-    @Test(expected = ParserException.class)
-    public void parse_mediaStoresMediaRule_throwsException() {
+    @Test
+    public void parse_mediaStoresMediaRule_supported() {
         final Log logger = Mockito.mock(Log.class);
         final SteadyStateParser parser = new SteadyStateParser(logger);
 
-        parser.parse("@media all { @media all { } }", options);
+        final StyleSheet stylesheet = parser.parse("@media all { @media all { } }", options);
+
+        final ComplexRule media = (ComplexRule) stylesheet.getRules().get(0);
+        Assert.assertEquals("Stylesheet nested rule class.", ComplexRule.class, media.getRules().get(0).getClass());
     }
 
     @Test
@@ -78,5 +82,18 @@ public final class Css30CompatibilityTest {
 
         final ComplexRule media = (ComplexRule) stylesheet.getRules().get(0);
         Assert.assertEquals("Stylesheet nested rule class.", UnknownRule.class, media.getRules().get(0).getClass());
+    }
+
+    @Test(expected = ParserException.class)
+    public void parse_starHackIsDisabled_throwsException() {
+        new StarHackTester(STANDARD, false).parse();
+    }
+
+    @Test
+    public void parse_starHackIsEnabled_returnsStyleSheet() {
+        final StarHackTester tester = new StarHackTester(STANDARD, true);
+        final StyleSheet stylesheet = tester.parse();
+        Assert.assertNotNull("Stylesheet.", stylesheet);
+        tester.verify(stylesheet);
     }
 }