+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # CSS Splitter Maven Plugin Changelog
 
+## 1.2.0
+Features:
+* Added support for nested `@media` rules in CSS 3.0
+* Added support for "star hack" in CSS 2.1 and 3.0 (see [starHackAllowed](http://css-splitter-maven-plugin.projects.gabrys.biz/1.2.0/split-mojo.html#starHackAllowed) parameter)
+
+Dependencies:
+* Upgraded [CSS Parser](http://cssparser.sourceforge.net/) from 0.9.20 to 0.9.21
+
+[See documentation](http://css-splitter-maven-plugin.projects.gabrys.biz/1.2.0/)
+
 ## 1.1.1
 Dependencies:
 * Upgraded [Maven Plugin Utils](http://maven-plugin-utils.projects.gabrys.biz/) from [1.3.0](http://maven-plugin-utils.projects.gabrys.biz/1.3.0/) to [1.4.0](http://maven-plugin-utils.projects.gabrys.biz/1.4.0/)